Move project to src folder
